reorganize taylor files
